final for review: - move open detailed film card event listener to component - not working; - move render and createElement from util.js to render.js; - refactor render in render.js to work with elements and components; - add remove in render.s to remove components;
