+++ src/util/util.js
@@ -1,35 +1,4 @@
 import dayjs from 'dayjs';
-import { RenderPosition } from './const.js';
-
-/**
- * Вместо шаблонной строки принимает элемент
- * @param {DOM element} container  - DOM-узел, куда добавляем элемент
- * @param {DOM element} element - ссылка на DOM-узел, т.е. элемент, котоырй получился в рез-те обработки разметки и который нужно добавить в контейнер
- * @param {constant} place - местоположение в контейнетре: в начале или в конце
- */
-export const render = (container, element, place) => {
-  switch (place) {
-    case RenderPosition.AFTERBEGIN:
-      container.prepend(element);
-      break;
-    case RenderPosition.BEFOREEND:
-      container.append(element);
-      break;
-  }
-};
-
-
-/**
- * Функция создания DOM-элемента на основании разметки
- * @param {string} template - принимает шаблонную строку. Строка должна иметь общую обертку.
- * @returns DOM-элемент
- */
-export const createSiteElement = (template) => {
-  const newElement = document.createElement('div');
-  newElement.innerHTML = template;
-
-  return newElement.firstElementChild;
-};
 
 const formatCommentDate = (date) => {
   return dayjs(date).format('YYYY/MM/DD HH:mm ');
