+++ src/view/abstract.js
@@ -1,4 +1,4 @@
-import { createSiteElement } from '../util/util.js';
+import { createSiteElement } from '../util/render.js';
 
 export default class Abstract {
   constructor() {
